requirements.txt
google-cloud-aiplatform==1.133.0
google-cloud-storage==3.1.1
google-genai==1.58.0
GitPython==3.1.46
google-adk==1.22.1
mcp==1.26.0
requests==2.32.5
beautifulsoup4==4.14.3